Refactor: Localize language names in `LanguageProvider`
This commit refactors the `getLanguageName` method in `LanguageProvider` to use `AppLocalizations` for displaying language names. This ensures that language names are localized according to the current application locale.

The `BuildContext` is now passed to `getLanguageName` to access `AppLocalizations`.

Additionally, new localization keys for "spain" and "german" have been added to `intl_en.arb`.