feat: Add centralized mock generation setup
This commit introduces a new file, `test/mocks.dart`, which centralizes the generation of mock objects for testing purposes using the `mockito` package.

The `@GenerateMocks` annotation is used to specify the classes for which mocks should be generated. These include:

*   **Services:** `AuthService`, `FirestoreService`
*   **Firebase & Google:** `FirebaseAuth`, `GoogleSignIn`, `User`, `UserCredential`, `GoogleSignInAccount`, `GoogleSignInAuthentication`, `AdditionalUserInfo`

This setup allows for easier management and generation of mocks by running `flutter pub run build_runner build --delete-conflicting-outputs`.